+++ packages/react-wrap-with/__tests__/types/extract.optional.empty.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 
-import { ExtractProp, type HowOf, wrapWith } from '../../src/index';
+import { Extract, type HowOf, wrapWith } from '../../src/index';
 
 import type { PropsWithChildren } from 'react';
 
@@ -9,7 +9,7 @@ const Header = ({ children, className }: PropsWithChildren<{ className?: string 
 );
 
 // Because "className" is extracted, it should be okay to pass {} for initialProps.
-const Component = wrapWith(Header, { className: ExtractProp } satisfies HowOf<typeof Header>)(() => <div />);
+const Component = wrapWith(Header, { className: Extract } satisfies HowOf<typeof Header>)(() => <div />);
 
 // It is okay to render without "className" prop because "className" prop is optional.
 <Component />;