+++ packages/react-wrap-with/__tests__/types/extract.required.missing.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 
-import { ExtractProp, type HowOf, wrapWith } from '../../src/index';
+import { Extract, type HowOf, wrapWith } from '../../src/index';
 
 import type { PropsWithChildren } from 'react';
 
@@ -8,7 +8,7 @@ const Header = ({ children, className }: PropsWithChildren<{ className: string }
   <h1 className={className}>{children}</h1>
 );
 
-const Component = wrapWith(Header, { className: ExtractProp } satisfies HowOf<typeof Header>)(() => <div />);
+const Component = wrapWith(Header, { className: Extract } satisfies HowOf<typeof Header>)(() => <div />);
 
 // "className" prop is mapped and is required.