+++ packages/react-wrap-with/__tests__/types/how.extract.wrongType.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 
-import { type HowOf, SpyProp } from '../../src/index';
+import { type HowOf, Spy } from '../../src/index';
 
 import type { PropsWithChildren } from 'react';
 
@@ -9,6 +9,6 @@ const Header = ({ children, className }: PropsWithChildren<{ className: string }
 );
 
 ({
-  // @ts-expect-error Type 'typeof SpyProp' is not assignable to type 'typeof ExtractProp'.
-  className: SpyProp
+  // @ts-expect-error Type 'typeof Spy' is not assignable to type 'typeof Extract'.
+  className: Spy
 }) satisfies HowOf<typeof Header, 'className'>;