+++ packages/react-wrap-with/__tests__/types/how.extraneous.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 
-import { ExtractProp, type HowOf } from '../../src/index';
+import { Extract, type HowOf } from '../../src/index';
 
 import type { PropsWithChildren } from 'react';
 
@@ -10,5 +10,5 @@ const Header = ({ children, className }: PropsWithChildren<{ className: string }
 
 ({
   // @ts-expect-error Object literal may only specify known properties, and 'value' does not exist in type 'HowOf<({ children, className }: PropsWithChildren<{ className: string; }>) => Element, never, "className">'.
-  value: ExtractProp
+  value: Extract
 }) satisfies HowOf<typeof Header, never, 'className'>;