+++ packages/react-wrap-with/__tests__/types/spy.noChildren.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 
-import { type HowOf, SpyProp } from '../../src/index';
+import { type HowOf, Spy } from '../../src/index';
 
 import type { ReactNode } from 'react';
 
@@ -9,4 +9,4 @@ const Header = (_: { children?: ReactNode | undefined }) => <h1>Hello, World!</h
 // Initial props must not have "children".
 
 // @ts-expect-error Type 'symbol' is not assignable to type 'undefined'.
-({ children: SpyProp }) satisfies HowOf<typeof Header>;
+({ children: Spy }) satisfies HowOf<typeof Header>;