+++ packages/react-wrap-with/__tests__/types/spy.wrongType.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 
-import { type HowOf, SpyProp, wrapWith } from '../../src/index';
+import { type HowOf, Spy, wrapWith } from '../../src/index';
 
 import type { PropsWithChildren } from 'react';
 
@@ -8,7 +8,7 @@ const Header = ({ children, className }: PropsWithChildren<{ className: string }
   <h1 className={className}>{children}</h1>
 );
 
-const Component = wrapWith(Header, { className: SpyProp } satisfies HowOf<typeof Header>)(() => <div />);
+const Component = wrapWith(Header, { className: Spy } satisfies HowOf<typeof Header>)(() => <div />);
 
 // @ts-expect-error Type 'boolean' is not assignable to type 'string'.
 <Component className={true} />;