+++ packages/react-wrap-with/__tests__/types/spy.wrongType2.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 
-import { type HowOf, SpyProp, wrapWith } from '../../src/index';
+import { type HowOf, Spy, wrapWith } from '../../src/index';
 
 import type { PropsWithChildren } from 'react';
 
@@ -9,4 +9,4 @@ const Header = ({ children, className }: PropsWithChildren<{ className: string }
 );
 
 // @ts-expect-error Type 'string' is not assignable to type 'number'.
-wrapWith(Header, { className: SpyProp } satisfies HowOf<typeof Header>)((_: { className: number }) => <div />);
+wrapWith(Header, { className: Spy } satisfies HowOf<typeof Header>)((_: { className: number }) => <div />);