+++ packages/react-wrap-with/src/Extract.ts
@@ -1,0 +1,3 @@
+const Extract = Symbol.for('react-wrap-with/Extract');
+
+export default Extract;