+++ packages/react-wrap-with/src/ExtractProp.ts
@@ -1,3 +1,0 @@
-const ExtractProp = Symbol.for('react-wrap-with/ExtractProp');
-
-export default ExtractProp;