+++ packages/react-wrap-with/src/Spy.ts
@@ -1,0 +1,3 @@
+const Spy = Symbol.for('react-wrap-with/Spy');
+
+export default Spy;