+++ packages/react-wrap-with/src/SpyProp.ts
@@ -1,3 +1,0 @@
-const SpyProp = Symbol.for('react-wrap-with/SpyProp');
-
-export default SpyProp;