+++ packages/react-wrap-with/src/private/wrapWith.extract.noContainer.spec.tsx
@@ -5,7 +5,7 @@ import { render } from '@testing-library/react';
 import PropTypes from 'prop-types';
 import React from 'react';
 
-import ExtractProp from '../ExtractProp';
+import Extract from '../Extract';
 import wrapWith from './wrapWith';
 
 const Hello = ({ className }: { className?: string }) => <h1 className={className}>Hello, World!</h1>;
@@ -14,7 +14,7 @@ Hello.propTypes = { className: PropTypes.string };
 
 test('extract props without a container', () => {
   // GIVEN: Wrapping <Hello> with false and extract "className" prop.
-  const BlinkingHello = wrapWith(false, { className: ExtractProp })(Hello);
+  const BlinkingHello = wrapWith(false, { className: Extract })(Hello);
 
   // WHEN: Render.
   const result = render(<BlinkingHello className="blink" />);